requirements.txt
wheel
Django==2.1.7
django-modeltranslation
django-grappelli
python-telegram-bot==12.0.0b1
django-jsonview
django-celery==3.2.2
redis==2.10.6
social-auth-app-django
bs4
django-crispy-forms
py-trello
django-tinymce
django-filebrowser
Pillow
tornado==4.5.3
tornado-redis
psycopg2
psycopg2-binary